+++ package.json
@@ -42,7 +42,6 @@
     "tiptap": "^1.26.6",
     "tiptap-extensions": "^1.28.6",
     "vue": "^2.6.11",
-    "vue-authenticate": "^1.4.1",
     "vue-chartjs": "^3.5.0",
     "vue-check-view": "^0.3.0",
     "vue-gtag": "^1.6.2",
