+++ src/app/plugins/index.js
@@ -7,6 +7,5 @@ export * from './vue-apollo'
 export * from './vue-check-view'
 export * from './vue-gtag'
 export * from './vue-social-sharing'
-export * from './vue-authenticate'
 export * from './network-status'
 export * from './random-id'
