+++ src/app/views/river-detail/accidents-tab/accidents-tab.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="accidents-tab">
-    <layout :name="!loading && !error ? 'layout-two-thirds' : 'layout-full-width'">
+    <layout name="layout-full-width">
       <template #main>
         <template v-if="loading">
           <utility-block
@@ -10,6 +10,16 @@
           />
         </template>
         <template v-else-if="accidents">
+          <page-description
+            description="If someone gets hurt on a river, or you read about a whitewater-related injury, please report it to American Whitewater. Don't worry about multiple submissions from other witnesses, as our safety editors will turn multiple witness reports into a single unified accident report."
+          >
+            <cv-button
+              kind="secondary"
+              size="small"
+            >
+              Submit Report
+            </cv-button>
+          </page-description>
           <div
             class="bx--data-table-container mb-lg"
           >
@@ -95,12 +105,13 @@ import { accidentsActions } from '../shared/state'
 import { accidentDetailActions } from '@/app/views/accident-database/shared/state'
 import UtilityBlock from '@/app/global/components/utility-block/utility-block'
 import { Layout } from '@/app/global/layout'
-
+import { PageDescription } from '@/app/global/components'
 export default {
   name: 'accidents-tab',
   components: {
     UtilityBlock,
-    Layout
+    Layout,
+    PageDescription
   },
   computed: {
     ...mapState({
