+++ src/app/views/river-detail/flow-tab/flow-tab.vue
@@ -132,7 +132,7 @@ import { GageChart, GageReadings, GageChartControls } from './components'
 import { GageChartConfig } from './utils/gage-chart-config'
 import { Layout } from '@/app/global/layout'
 import { mapState } from 'vuex'
-import { readingsActions } from '../shared/state'
+import { readingsActions, reachGagesActions } from '../shared/state'
 import UtilityBlock from '@/app/global/components/utility-block/utility-block'
 import { checkWindow } from '@/app/global/mixins'
 
@@ -170,6 +170,9 @@ export default {
       gagesLoading: state => state.riverDetailState.reachGagesData.loading,
       gagesError: state => state.riverDetailState.reachGagesData.error
     }),
+    riverId () {
+      return this.$route.params.id
+    },
     /**
      * vue-chartjs requires data to be formatted this way
      * @reference https://vue-chartjs.org/guide/#example
@@ -250,6 +253,9 @@ export default {
         this.chartConfig.scales.xAxes[0].time.unit = format
       }
     }
+  },
+  created () {
+    this.$store.dispatch(reachGagesActions.FETCH_GAGES, this.riverId)
   }
 }
 </script>
